+++ presentation/src/main/java/jjh/movie/app/presentation/ui/home/adapter/BoxOfficeAdapter.kt
@@ -1,4 +1,4 @@
-package jjh.movie.app.presentation.ui
+package jjh.movie.app.presentation.ui.home.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import jjh.movie.app.presentation.databinding.ItemBoxOfficeBinding
+import jjh.movie.app.presentation.ui.home.viewholder.BoxOfficeViewHolder
 
 class BoxOfficeAdapter : ListAdapter<BoxOffice, ViewHolder>(boxOfficeDiffCallback) {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
