+++ presentation/src/main/res/drawable/shape_round_square_10.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
-    <solid android:color="@color/gray" />
     <corners android:radius="10dp" />
 </shape>
