+++ presentation/src/main/res/layout/item_ranking.xml
@@ -9,7 +9,8 @@
         android:id="@+id/iv_movie_poster"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@drawable/shape_ranking" />
+        android:background="@drawable/shape_round_square_10"
+        android:src="@color/gray" />
 
     <TextView
         android:id="@+id/tv_ranking"
