+++ presentation/src/main/res/navigation/bottom_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/home_fragment"
-        android:name="jjh.movie.app.presentation.ui.HomeFragment"
+        android:name="jjh.movie.app.presentation.ui.home.HomeFragment"
         android:label="OnboardingFragment"
         tools:layout="@layout/fragment_home" />
